move snappy.PackageFile -> pkg.File
